ConfigurationGathering: Disable CLI section
For now all the settings of the `cli` section is copied to the
`default` section and disabling it till the `default` section is
removed from coala.

This is to prevent the breakage of cli args options like `-f` and `-m`
till the time everyone has migrated to the new scheme of not
having a default section.